Fix "Implicit conversion from float 0.02 to int loses precision in"

+++ maintenance/resetWikiCaches.php
@@ -17,7 +17,7 @@ class ResetWikiCaches extends Maintenance {
 		$cWJ = new CreateWikiJson( $config->get( 'DBname' ) );
 		$cWJ->resetWiki();
 
-		sleep( 0.02 );
+		sleep( (int)0.02 );
 	}
 }
